requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
coverage==7.5.2
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==25.2.0
inflection==0.5.1
iniconfig==2.0.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.1.19
pudb==2024.1
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-pudb==0.7.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
rules==3.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urwid==2.6.12
urwid_readline==0.14
wcwidth==0.2.13
